Implement com.dsi.ant hidl client.
Changes the hal service to be a hidl client instead of using
libantradio.so. This is part of the changes needed to support
project treble.

Some config variables have also been moved into an xml
resource. This allows vendors to customize behaviour using an
overlay rather than modifying the hal service source.